\name{coerce}
\alias{as.data.frame.rtv}
\alias{as.Date.rtv}
\alias{as.POSIXlt.rtv}
\alias{as.POSIXct.rtv}
\title{coerce rtv objects to other objects}
\description{These functions should be self-explanatory.}
\usage{
\method{as.data.frame}{rtv}(x, \dots)
\method{as.Date}{rtv}(x, \dots)
\method{as.POSIXlt}{rtv}(x, \dots)
\method{as.POSIXct}{rtv}(x, \dots)
}
\arguments{
\item{x}{.}
\item{\dots}{.}
}



