\name{is.rtv}
\alias{is.rtv}
\alias{is.drtv}
\alias{is.crtv}
\title{Testing for Random Time Variable Objects}
\description{Functions to test if an object is of class rtv, drtv or crtv, noting that classes drtv and crtv extend class rtv.}
\usage{
is.rtv(x)
is.drtv(x)
is.crtv(x)
}
\arguments{\item{x}{Object to be checked.}}

