\name{math}
\alias{sort.rtv}
\alias{sample.rtv}
\alias{length.drtv}
\alias{length.crtv}
\alias{order.drtv}
\alias{order.crtv}
\alias{c.drtv}
\alias{c.crtv}
\alias{rep.drtv}
\alias{rep.crtv}
\alias{[.drtv}
\alias{[.crtv}
\alias{mean.crtv}
\alias{min.crtv}
\alias{max.crtv}
\alias{range.crtv}
\alias{round.crtv}
\alias{floor.crtv}
\alias{ceiling.crtv}
\alias{+.crtv}
\alias{-.crtv}
\alias{sample}
\alias{sample.default}
\title{standard math operations}
\description{todo}
\usage{
\method{sort}{rtv}(x, \dots)
\method{sample}{rtv}(x, \dots)
\method{length}{drtv}(x, \dots)
\method{length}{crtv}(x, \dots)
\method{order}{drtv}(x, by.unit = getOption("rtv.unit"),
	by.cp = FALSE, \dots)
\method{order}{crtv}(x, by.unit = x$unit, by.cp = FALSE, \dots)
\method{c}{drtv}(x, \dots)
\method{c}{crtv}(x, \dots)
\method{rep}{drtv}(x, times, \dots)
\method{rep}{crtv}(x, times, \dots)
\method{[}{drtv}(x, i)
\method{[}{crtv}(x, i)
\method{mean}{crtv}(x, \dots)
\method{min}{crtv}(x, \dots)
\method{max}{crtv}(x, \dots)
\method{round}{crtv}(x, \dots)
\method{floor}{crtv}(x, \dots)
\method{ceiling}{crtv}(x, \dots)
\method{range}{crtv}(x, \dots)
\method{+}{crtv}(a, b)
\method{-}{crtv}(a, b)
sample (\dots)
\method{sample}{default}(x, \dots)
}
\arguments{
\item{x}{.}
\item{i}{.}
\item{by.unit}{.}
\item{by.cp}{.}
\item{times}{.}
\item{a}{.}
\item{b}{.}
\item{\dots}{.}
}

