\name{rtvo}
\alias{rtvo.reset}
\alias{rtvo.format}
\alias{rtvo.date}
\title{rtv options}
\description{Functions to set rtv options. The function rtvo.reset, resets all rtv options to their defaults.}
\usage{
rtvo.reset ()
rtvo.format (format=FALSE)
rtvo.date (date=TRUE)
}
\arguments{
\item{format}{.}
\item{date}{.}
}

