\name{timeaxis}
\alias{timeaxis}
\title{Create Formatted Time Axis for Custom Plots}
\description{A formatted axis for custom plots. Wraps the side function.}
\usage{timeaxis(side, x, n = 5, \dots)}
\arguments{
\item{side}{Side of the axis. See axis, for more info.}
\item{x}{A crtv object. Must be crtv, not drtv.}
\item{n}{Number of ticks.}
\item{\dots}{Arguments for explicit.format.}
}

