\name{timebox}
\alias{timebox}
\title{Formatted Time Box for Custom Plots}
\description{A convenience function. This calls box, creates a timeaxis, and potentially another axis. See the functions timeaxis, box or axis for more information.}
\usage{timebox (side, x, \dots)}
\arguments{
\item{side}{The side, can have length one or two. The value side [1] indicates which side has the timeaxis. The value side [2] if exists, is a regular axis.}
\item{x}{A crtv object. Must be crtv, not drtv.}
\item{\dots}{Arguments for timestring.}
}

