\name{timeplot}
\alias{plot.rtv}
\alias{timeplot.ecdf}
\alias{timeplot.dot}
\alias{timeplot.group}
\alias{timeplot.line}
\title{Time Plots}
\description{Functions for creating plots with a formatted time axis. Note these are prototypical functions, and may be changed substantially in future releases.}
\usage{
plot.rtv (x, y, \dots)
timeplot.ecdf (x, cu=NULL, \dots)
timeplot.dot (x, \dots)
timeplot.group (g, x, cu=NULL, jitter=TRUE, \dots)
timeplot.line (x, y, cu=NULL, sort=TRUE, \dots)
}
\arguments{
\item{x}{}
\item{y}{}
\item{cu}{}
\item{jitter}{}
\item{sort}{}
\item{g}{}
\item{\dots}{}
}

