\name{timeseq}
\alias{timeseq}
\title{A Sequence of Time Points}
\description{A sequence of time points represented by a crtv object.}
\usage{timeseq(x1, x2, n, \dots)}
\arguments{
\item{x1}{An rtv object either of length one, giving the starting point, or, of length two, giving both the starting and ending points.}
\item{x2}{A scalar rtv object giving the ending point. Not required if x1 gives both points.}
\item{n}{Number of points in the sequence.}
\item{\dots}{Arguments for crtv, e.g. unit.}
}

