\name{timesweep}
\alias{timesweep}
\title{Applying Functions to rtv Objects}
\description{This is a convenience function for applying typical mathematical functions to rtv objects. Many rtv methods use this function. See the source for mean.rtv for an example.}
\usage{timesweep (f, x, \dots)}
\arguments{
\item{f}{A function to apply.}
\item{x}{An rtv object.}
\item{\dots}{Arguments for the child function f, e.g. na.rm=TRUE.}
}

