\name{when}
\alias{when}
\title{Map a time.frame to an rtv Object}
\description{This function is designed for time.frame objects. It will return the first rtv in the time.frame. It can also be called with rtv objects (in which case, the same rtv object is returned), or on regular data.frame objects. If the object does not contain an rtv object, then NULL is returned.}
\usage{when (x)}
\arguments{
\item{x}{A time.frame object, or other supported object.} 
}

