\name{format.rtv}
\alias{format.rtv}
\title{Formatting rtv Objects}
\description{A format method for rtv objects. Note that by default nothing happens. There is an option to change this behaviour, refer to the package vignette. Alternatively use timestring.}
\usage{\method{format}{rtv}(x, \dots)}
\arguments{
\item{x}{An rtv object.}
\item{\dots}{Arguments for timestring.}
}

