\name{timeaxis}
\alias{timeaxis}
\title{Formatted Time Axis for Custom Plots}
\description{This creates a formatted time axis for custom plots. This function calls both axis and timestring. Hence to understand how to call this function, refer to axis or timestring.}
\usage{timeaxis (side, x, n=5, date=getOption ("rtv.plot.date"), \dots)}
\arguments{
\item{side}{Side of the axis. See axis, for more info.}
\item{x}{An rtv object.}
\item{n}{Number of ticks.}
\item{date}{Date or date-tod format.}
\item{\dots}{Arguments for timestring.}
}

