\name{timeseq}
\alias{timeseq}
\title{A Sequence of Time Values}
\description{A (deterministic and uniformly spaced) sequence of time values represented by a crtv object.}
\usage{timeseq (x1, x2, n, \dots)}
\arguments{
\item{x1}{An rtv object of either, length one, giving the starting value, or length two, giving both the starting and ending values.}
\item{x2}{A scalar rtv object giving the ending value (if x1 is of length one).}
\item{n}{Length of the sequence.}
\item{\dots}{Arguments for the crtv constructor, e.g. unit.}
}

